smart-bookmarks-django/pyproject.toml
[tool.poetry]
name = "smart-bookmarks-django"
version = "0.1.0"
description = ""
authors = ["Tomasz Piwoński <tpiwonski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "2.2.10"
gunicorn = "20.0.4"
psycopg2-binary = "2.8.4"
celery = "4.4.0"
redis = "3.3.11"
django-redis = "4.11.0"
flower = "0.9.3"
selenium = "3.141.0"
python-dotenv = "0.11.0"
elasticsearch-dsl = "7.1.0"
django-widget-tweaks = "1.4.5"

[tool.poetry.dev-dependencies]
isort = "^4.3.21"
black = "^19.10b0"
mypy = "^0.761"
pytest = "5.3.5"
pytest-django = "3.8.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
